requirements.txt
absl-py==0.11.0
dill==0.3.3
Keras-Preprocessing==1.1.2
nlp==0.4.0
numpy==1.18.5
pandas==1.1.4
pytorch-lightning==1.0.6
sh==1.14.1
tensorboard==2.4.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tokenizers==0.9.3
torch==1.7.0
transformers==3.5.1
